+++ navigate/output/plots/port_bunker_price.py
@@ -22,9 +22,9 @@ def plot_port_bunker_price(manager, directory):
     ports = manager.nodes.ports
 
     for port_name, port in ports.items():
+        bunker_price = port.profile.get_bunker_price()
         fuel_costs = {
-            fuel.name: port.profile.get_bunker_price(fuel.name)
-            / fuel.lower_heating_value.get()
+            fuel.name: bunker_price[fuel.name] / fuel.lower_heating_value.get()
             for fuel in fuels.values()
         }